#49 Pelinäkymä siirretty App.js tiedostosta omaksi näkymäksi
